// types/tokenfactory/amp-compounder/compound_proxy/execute_msg.d.ts
export type ExecuteMsg =
  | {
      compound: {
        /**
         * LP into which the assets should be compounded into
         */
        lp_token: string;
        /**
         * Skip optimal swap
         */
        no_swap?: boolean | null;
        /**
         * Receiver address for LP token
         */
        receiver?: string | null;
        /**
         * List of reward asset send to compound
         */
        rewards: Asset[];
        /**
         * slippage tolerance when providing LP
         */
        slippage_tolerance?: Decimal | null;
      };
    }
  | {
      multi_swap: {
        /**
         * List of reward asset send to compound
         */
        assets: Asset[];
        /**
         * LP into which the assets should be compounded into
         */
        into: AssetInfo;
        /**
         * Receiver address for LP token
         */
        receiver?: string | null;
      };
    }
  | {
      propose_new_owner: {
        /**
         * The validity period of the proposal to change the owner
         */
        expires_in: number;
        /**
         * The newly proposed owner
         */
        owner: string;
      };
    }
  | {
      update_config: {
        default_max_spread?: number | null;
        delete_lps?: string[] | null;
        delete_routes?: RouteDelete[] | null;
        factory?: string | null;
        insert_routes?: RouteInit[] | null;
        remove_factory?: boolean | null;
        upsert_lps?: LpInit[] | null;
      };
    }
  | {
      drop_ownership_proposal: {};
    }
  | {
      claim_ownership: {};
    }
  | {
      callback: CallbackMsg;
    };
// ...
export type Uint128 = string;
/**
 * This enum describes available Token types. ## Examples ``` # use cosmwasm_std::Addr; # use astroport::asset::AssetInfo::{NativeToken, Token}; Token { contract_addr: Addr::unchecked("stake...") }; NativeToken { denom: String::from("uluna") }; ```
 */
export type AssetInfo =
  | {
      token: {
        contract_addr: Addr;
      };
    }
  | {
      native_token: {
        denom: string;
      };
    };
// ...
export type Addr = string;
/**
 * A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0
 *
 * The greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)
 */
export type Decimal = string;
export type RouteInit =
  | {
      path: {
        route: AssetInfo[];
        router: string;
        router_type: RouterType;
      };
    }
  | {
      pair_proxy: {
        pair_contract: string;
        /**
         * when specified, a pair can be defined as a single direction
         */
        single_direction_from?: AssetInfo | null;
      };
    }
  | {
      unwrap: {
        contract: string;
        from: AssetInfo;
        to: AssetInfo;
      };
    };
export type RouterType =
  | ("astro_swap" | "terra_swap" | "token_swap")
  | {
      t_f_m: {
        route: [string, Addr][];
        [k: string]: unknown;
      };
    };
export type LpType = "astroport" | "white_whale";
/**
 * This structure describes the callback messages of the contract.
 */
export type CallbackMsg =
  | {
      optimal_swap: {
        lp_token: string;
      };
    }
  | {
      provide_liquidity: {
        lp_token: string;
        prev_balances: Asset[];
        receiver: string;
        slippage_tolerance?: Decimal | null;
      };
    }
  | {
      send_swap_result: {
        receiver: string;
        token: AssetInfo;
      };
    };

/**
 * This enum describes a Terra asset (native or CW20).
 */
export interface Asset {
  /**
   * A token amount
   */
  amount: Uint128;
  /**
   * Information about an asset stored in a [`AssetInfo`] struct
   */
  info: AssetInfo;
}
export interface RouteDelete {
  both?: boolean | null;
  from: AssetInfo;
  to: AssetInfo;
}
export interface LpInit {
  /**
   * The swap commission
   */
  commission_bps: number;
  lp_type?: LpType | null;
  /**
   * The pair info
   */
  pair_contract: string;
  /**
   * The slippage tolerance when providing liquidity
   */
  slippage_tolerance: Decimal;
  /**
   * Token used for providing liquidity
   */
  wanted_token: AssetInfo;
}